Demo decorator approach
Meh, not worth it. Making it actually work for each widget would require
passing in a value reconciliation function tailored for each, and with
the amount of checking for arguments existing required to handle the
functions receiving a very generic set of arguments in practice, it
doesn't seem to add any value over writing it all out.